requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
coverage==6.4.3
django==4.0.6
django-extensions==3.2.0
django-ninja==0.19.1
django-ninja-jwt==5.2.0
djangorestframework==3.13.1
idna==3.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pydantic==1.9.1
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
typing-extensions==4.3.0
urllib3==1.26.11
